変更 - Iconではなく、Imageのスタイル指定だったため、内容を移設
