+++ src/component/Element/Image/index.ts
@@ -1,0 +1,3 @@
+import { IconElement } from './Icon';
+
+export { IconElement };
